+++ sdk/storage/Azure.Storage.Files.Shares/tests/ShareClientTestFixtureAttribute.cs
@@ -51,8 +51,8 @@ namespace Azure.Storage.Files.Shares.Tests
                 },
                 additionalParameters: additionalParameters)
         {
-            RecordingServiceVersion = StorageVersionExtensions.MaxVersion;
-            LiveServiceVersions = new object[] { StorageVersionExtensions.LatestVersion, };
+            RecordingServiceVersion = ShareClientOptions.ServiceVersion.V2026_10_06;
+            LiveServiceVersions = new object[] { ShareClientOptions.ServiceVersion.V2026_10_06, };
         }
     }
 }